pass typeMap to factory
this is useful in a plugin scenario, where both the plugin and the main application link statically against a library which includes these headers. Both will have distinct instances of the static variables, e.g. the typeRegistry. If a plugin is to define a reflected type and the main application uses the typeRegistry for lookup, it must pass its instance of the typeRegistry to the plugin